重构map代码
