+++ .vscode/settings.json
@@ -14,6 +14,10 @@
         "functional": "c",
         "tuple": "c",
         "type_traits": "c",
-        "utility": "c"
+        "utility": "c",
+        "map_ini_cfg.h": "c",
+        "map_data.h": "c",
+        "ini_cfg.h": "c",
+        "errcode.h": "c"
     }
 }
