+++ src/map/map.h
@@ -1,6 +1,6 @@
 /**
  * @file map.h
- * @author your name (you@domain.com)
+ * @author zhangmeng (847305142@qq.com)
  * @brief 地图功能对外接口汇总
  * @version 0.1
  * @date 2023-07-13
